commented overtaking

+++ group-8/sources/OpenDaVINCI-msv/apps/overtaker/include/Overtaker.h
@@ -76,7 +76,7 @@ namespace msv {
             double m_steering;
             int m_sensor;
             double m_carLength;
-            int m_speed;
+            double m_speed;
 
     };
